Move JS files used only on main map page to their own directory
Also adjust webpack code to create aliases for all folders in `js/src`, not just `lib`
