+++ opentreemap/treemap/js/src/mapPage/addResourceMode.js
@@ -4,7 +4,7 @@ var $ = require('jquery'),
     _ = require('lodash'),
     L = require('leaflet'),
     U = require('treemap/lib/utility.js'),
-    addMapFeature = require('treemap/lib/addMapFeature.js'),
+    addMapFeature = require('treemap/mapPage/addMapFeature.js'),
     polylineEditor = require('treemap/lib/polylineEditor.js'),
     config = require('treemap/lib/config.js'),
     reverse = require('reverse');
