+++ opentreemap/treemap/js/src/mapPage/editTreeDetailsMode.js
