+++ opentreemap/treemap/js/src/mapPage/modes.js
@@ -10,10 +10,10 @@ var $                   = require('jquery'),
     reverse             = require('reverse'),
     config              = require('treemap/lib/config.js'),
     U                   = require('treemap/lib/utility.js'),
-    browseTreesMode     = require('treemap/lib/browseTreesMode.js'),
-    addTreeMode         = require('treemap/lib/addTreeMode.js'),
-    editTreeDetailsMode = require('treemap/lib/editTreeDetailsMode.js'),
-    addResourceMode     = require('treemap/lib/addResourceMode.js'),
+    browseTreesMode     = require('treemap/mapPage/browseTreesMode.js'),
+    addTreeMode         = require('treemap/mapPage/addTreeMode.js'),
+    editTreeDetailsMode = require('treemap/mapPage/editTreeDetailsMode.js'),
+    addResourceMode     = require('treemap/mapPage/addResourceMode.js'),
     inlineEditForm      = require('treemap/lib/inlineEditForm.js'),
     urlState            = require('treemap/lib/urlState.js'),
     plotMarker          = require('treemap/lib/plotMarker.js'),
